+++ russenger/src/lib.rs
@@ -147,7 +147,7 @@ pub mod query;
 pub mod response_models;
 
 pub use core::{
-    app::App,
+    action::Router,
     services::{webhook_core, webhook_verify}, // core services
 };
 pub mod error {
@@ -162,6 +162,116 @@ pub use russenger_macro::action;
 
 use actix_files as fs;
 use actix_web::{web, App as ActixApp, HttpServer};
+
+use std::{collections::HashSet, sync::Arc};
+
+use crate::query::Query;
+use tokio::sync::Mutex;
+
+#[derive(Clone)]
+pub struct ActionLock {
+    pub locked_users: Arc<Mutex<HashSet<String>>>,
+}
+
+impl ActionLock {
+    pub async fn lock(&self, user: &str) -> bool {
+        let mut locked_user = self.locked_users.lock().await;
+        if !locked_user.contains(user) {
+            locked_user.insert(user.to_string());
+            true
+        } else {
+            false
+        }
+    }
+
+    pub async fn unlock(&self, user: &str) {
+        let mut locked_users = self.locked_users.lock().await;
+        locked_users.remove(user);
+    }
+}
+
+/// # App State
+///
+/// This module contains the `AppState` struct which is used to store the state of the application.
+///
+/// # Fields
+///
+/// * `query`: A `Query` that represents the query made by the user.
+#[derive(Clone)]
+pub struct App {
+    pub(crate) query: Query,
+    pub(crate) router: Arc<Mutex<Router>>,
+    pub(crate) tmp_router: Router,
+    pub(crate) action_lock: ActionLock,
+}
+
+impl App {
+    /// `init` is method to create new `App` instance. in russenger
+    pub async fn init() -> Result<Self> {
+        let query: Query = Query::new().await?;
+        Ok(Self {
+            query,
+            router: Arc::new(Mutex::new(Router::new())),
+            action_lock: ActionLock {
+                locked_users: Arc::new(Mutex::new(HashSet::new())),
+            },
+            tmp_router: Router::new(),
+        })
+    }
+
+    /// `attach` method allows you to add a group of predefined routes (actions) to the application's router.
+    ///
+    /// # Parameters
+    /// - `router`: A `Router` instance containing a set of routes to be attached.
+    ///
+    /// #  Examples
+    ///
+    /// ```rust
+    /// use russenger::prelude::*;
+    ///
+    /// #[action]
+    /// async fn index(res: Res, req: Req) -> Result<()> {
+    ///     res.send(TextModel::new(&req.user, "hello world")).await?;
+    ///     res.redirect("/next_action").await?;
+    ///     Ok(())
+    /// }
+    ///
+    /// #[action]
+    /// async fn next_action(res: Res, req: Req) -> Result<()> {
+    ///     let message: String = req.data.get_value();
+    ///     res.send(TextModel::new(&req.user, &message)).await?; // send the message to the user
+    ///     Ok(())
+    /// }
+    ///
+    /// pub fn group_actions() -> Router {
+    ///     let mut router = Router::new();
+    ///     router.add("/", index);
+    ///     router.add("/next_action", next_action);
+    ///     router
+    /// }
+    ///
+    /// #[russenger::main]
+    /// async fn main() -> Result<()> {
+    ///     let mut app = App::init().await?;
+    ///     app.attach(group_actions()).await; // Attach group actions to the application's router
+    ///     launch(app).await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    ///
+    /// This method is useful for organizing your application's routes into modular and reusable groups.
+    pub fn attach(&mut self, router: Router) -> Self {
+        self.tmp_router.extend(router);
+        self.clone()
+    }
+
+    pub async fn launch(&self) -> Result<()> {
+        dotenv().ok();
+        self.router.lock().await.extend(self.tmp_router.clone());
+        run_server(self.clone()).await?;
+        Ok(())
+    }
+}
 
 fn print_info(host: &str, port: u16) {
     let url = format!("http://{}:{}", host, port);
@@ -187,16 +297,5 @@ async fn run_server(app: App) -> Result<()> {
     .bind((host, port))?
     .run()
     .await?;
-    Ok(())
-}
-
-pub async fn launch(app_state: App) -> Result<()> {
-    dotenv().ok();
-    app_state
-        .router
-        .lock()
-        .await
-        .extend(app_state.tmp_router.clone());
-    run_server(app_state.clone()).await?;
     Ok(())
 }
